# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DARAI API",
    "description": "Локальная CRM совещаний. Контракт: docs/API_CONTRACT.md. Маршруты без префикса /api — его удаляет прокси frontend.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List Users",
        "operationId": "list_users_users_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserAdminOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "operationId": "create_user_users_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAdminOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Patch User",
        "operationId": "patch_user_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAdminOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmployeeBrief": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "fio": {
            "type": "string",
            "title": "Fio"
          },
          "position": {
            "type": "string",
            "title": "Position"
          },
          "department": {
            "type": "string",
            "title": "Department"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fio",
          "position",
          "department"
        ],
        "title": "EmployeeBrief"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginIn": {
        "properties": {
          "login": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Login"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "title": "LoginIn"
      },
      "LoginOut": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          }
        },
        "type": "object",
        "required": [
          "user",
          "csrf_token"
        ],
        "title": "LoginOut"
      },
...
      "Page_UserAdminOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserAdminOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[UserAdminOut]"
      },
...
      "UserAdminOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "login": {
            "type": "string",
            "title": "Login"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "secretary",
              "employee"
            ],
            "title": "Role"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "employee_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "employee_fio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Fio"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "login",
          "role",
          "active",
          "employee_id",
          "employee_fio",
          "created_at"
        ],
        "title": "UserAdminOut"
      },
      "UserCreate": {
        "properties": {
          "employee_id": {
            "type": "string",
            "format": "uuid",
            "title": "Employee Id"
          },
          "login": {
            "type": "string",
            "title": "Login"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 8,
            "title": "Password"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "secretary",
              "employee"
            ],
            "title": "Role"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "employee_id",
          "login",
          "password",
          "role"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "login": {
            "type": "string",
            "title": "Login"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "secretary",
              "employee"
            ],
            "title": "Role"
          },
          "employee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmployeeBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "login",
          "role",
          "employee"
        ],
        "title": "UserOut"
      },
      "UserPatch": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "admin",
                  "secretary",
                  "employee"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UserPatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}